+++ modules/internal/network/dynamicupstreams_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dynamicupstreams
+package network
 
 import (
 	"context"
